Move history out of generator
