--- src/main/java/tech/jhipster/lite/generator/history/domain/GeneratorHistoryData.java
@@ -1,5 +1,0 @@
-package tech.jhipster.lite.generator.history.domain;
-
-import java.util.List;
-
-public record GeneratorHistoryData(List<GeneratorHistoryValue> values) {}
