--- src/main/java/tech/jhipster/lite/generator/history/domain/GeneratorHistoryValue.java
@@ -1,5 +1,0 @@
-package tech.jhipster.lite.generator.history.domain;
-
-import java.time.Instant;
-
-public record GeneratorHistoryValue(String serviceId, Instant timestamp) {}
