+++ src/main/java/tech/jhipster/lite/history/domain/GeneratorHistoryData.java
@@ -1,0 +1,5 @@
+package tech.jhipster.lite.history.domain;
+
+import java.util.List;
+
+public record GeneratorHistoryData(List<GeneratorHistoryValue> values) {}
