+++ src/main/java/tech/jhipster/lite/history/domain/GeneratorHistoryRepository.java
@@ -1,0 +1,9 @@
+package tech.jhipster.lite.history.domain;
+
+import tech.jhipster.lite.module.domain.properties.JHipsterProjectFolder;
+
+public interface GeneratorHistoryRepository {
+  GeneratorHistoryData getHistoryData(JHipsterProjectFolder folder);
+
+  void addHistoryValue(HistoryProject project, GeneratorHistoryValue generatorHistoryValue);
+}
