+++ src/main/java/tech/jhipster/lite/history/domain/GeneratorHistoryValue.java
@@ -1,0 +1,5 @@
+package tech.jhipster.lite.history.domain;
+
+import java.time.Instant;
+
+public record GeneratorHistoryValue(String serviceId, Instant timestamp) {}
