+++ src/test/java/tech/jhipster/lite/history/domain/HistoryProjectsFixture.java
@@ -1,0 +1,12 @@
+package tech.jhipster.lite.history.domain;
+
+import tech.jhipster.lite.common.domain.FileUtils;
+
+public final class HistoryProjectsFixture {
+
+  private HistoryProjectsFixture() {}
+
+  public static HistoryProject tmpProject() {
+    return new HistoryProject(FileUtils.tmpDirForTest(), "\n");
+  }
+}
